Updated media library to display images
Modified various files to better handle and display images in the media
library. The changes include the creation of DOM elements for images,
adjustment of data retrieval methods for getting different image variants,
and the correction of queries related to retrieving files. This allows for a more
efficient and visually pleasing way to manage user-uploaded files in the system.